SpatialDataQC/requirements.txt
PyQt5>=5.15.0
matplotlib>=3.5.0
numpy>=1.21.0
Pillow>=9.0.0
PyYAML>=6.0
